backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
sse-starlette==2.2.1
python-multipart==0.0.20

# LangChain & LLM
langchain==0.3.14
langchain-openai==0.3.0
langchain-community==0.3.14
langgraph==0.2.60

# Database
sqlalchemy==2.0.36
pymysql==1.1.1
psycopg2-binary==2.9.10

# Data Processing
pandas==2.2.3
openpyxl==3.1.5
pdfplumber==0.11.4

# Vector Store (Few-shot RAG)
faiss-cpu==1.9.0
chromadb==0.5.23

# Visualization & Export
python-pptx==1.0.2
jinja2==3.1.5
markdown==3.7

# Utilities
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
aiofiles==24.1.0
